1.优化结构 2.增加presenter
