+++ app/src/main/java/com/example/basemvvm/ui/activity/BinderActivity.java
@@ -5,6 +5,7 @@ import com.example.basemvvm.R;
 import com.example.basemvvm.databinding.ActivityBinderBinding;
 import com.example.basemvvm.mvvm.viewModel.BinderVM;
 import com.wang.mvvmcore.base.activity.BaseSwipeBackLeftActivity;
+import com.wang.mvvmcore.base.baseViewModel.BaseToolbarVM;
 
 /**
  * author: wtg
@@ -25,6 +26,11 @@ public class BinderActivity extends BaseSwipeBackLeftActivity<ActivityBinderBind
     @Override
     protected int getLayoutRes() {
         return R.layout.activity_binder;
+    }
+
+    @Override
+    protected void bindOtherViewModel() {
+        binding.toolbar.setToolbarVM(new BaseToolbarVM(this));
     }
 
     @Override
