+++ app/src/main/res/layout/activity_binder.xml
@@ -3,24 +3,39 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
+
         <variable
             name="binderVM"
             type="com.example.basemvvm.mvvm.viewModel.BinderVM" />
     </data>
 
-    <com.scwang.smart.refresh.layout.SmartRefreshLayout
-        android:id="@+id/smartRefreshLayout"
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:srlEnableLoadMore="false"
-        onRefreshListener="@{binderVM.onRefreshListener}">
-        <com.scwang.smart.refresh.header.MaterialHeader
+        android:orientation="vertical">
+
+        <include
+            android:id="@+id/toolbar"
+            layout="@layout/view_toolbar" />
+
+        <com.scwang.smart.refresh.layout.SmartRefreshLayout
+            android:id="@+id/smartRefreshLayout"
+            onRefreshListener="@{binderVM.onRefreshListener}"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"/>
-        <androidx.recyclerview.widget.RecyclerView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            app:layoutManager="@{binderVM.linearLayoutManager}"
-            recycler_adapter="@{binderVM.baseMultiBinderAdapter}"/>
-    </com.scwang.smart.refresh.layout.SmartRefreshLayout>
+            android:layout_height="0dp"
+            android:layout_weight="1"
+            app:srlEnableLoadMore="false">
+
+            <com.scwang.smart.refresh.header.MaterialHeader
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content" />
+
+            <androidx.recyclerview.widget.RecyclerView
+                recycler_adapter="@{binderVM.baseMultiBinderAdapter}"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                app:layoutManager="@{binderVM.linearLayoutManager}" />
+        </com.scwang.smart.refresh.layout.SmartRefreshLayout>
+    </LinearLayout>
+
 </layout>
